Cleaned up code by factoring out progress report

+++ app/lib/detectors.scala
@@ -17,14 +17,14 @@ trait Detector {
   def run
 }
 
-class MongoStreamDetector(val key: String, val totalRecords: Option[Int] = None) extends Detector {
+class MongoStreamDetector(val key: String, val totalRecords: Option[Long] = None) extends Detector {
   def run {
     val rs = source.find().sort(Map(key -> 1)) map MongoUtils.toDocument
     Duplicates.windowedDetect(rs, collector, Model.windowSize, totalRecords=totalRecords)
   }
 }
 
-class MongoSortedHashDetector(val hashes: Int, val totalRecords: Option[Int] = None) extends Detector {
+class MongoSortedHashDetector(val hashes: Int, val totalRecords: Option[Long] = None) extends Detector {
   def run {
     val collector = new MongoDBCollector("candidates")
 
@@ -37,7 +37,7 @@ class MongoSortedHashDetector(val hashes: Int, val totalRecords: Option[Int] = N
   }
 }
 
-class MongoExternallySorted(val file: String, val totalRecords: Option[Int] = None) extends Detector {
+class MongoExternallySorted(val file: String, val totalRecords: Option[Long] = None) extends Detector {
   def run {
     val sortedHashes = new BufferedSource(new FileInputStream(file))
     val lines = sortedHashes.getLines
@@ -61,7 +61,7 @@ class MongoExternallySorted(val file: String, val totalRecords: Option[Int] = No
 }
 
 
-class PrefetchingMongoExternallySorted(val file: String, val totalRecords: Option[Int] = None) extends Detector {
+class PrefetchingMongoExternallySorted(val file: String, val totalRecords: Option[Long] = None) extends Detector {
   def run {
     val sortedHashes = new BufferedSource(new FileInputStream(file))
     val lines = sortedHashes.getLines
